+++ packages/ui-editor/src/components/primary-settings.tsx
@@ -1,0 +1,5 @@
+import { styled } from '../theme'
+
+export const PrimarySettings = styled.div({
+  marginTop: '15px'
+})